Fix Hikvision ITS plate recognition GBK decoding and plate color mapping.
Register CodePages encoding provider so GBK works on .NET 10, and cast byte plate color values to int before Enum.IsDefined.

+++ src/MaterialClient.Common/Services/Hikvision/HikvisionLprService.cs
@@ -813,7 +813,7 @@ public class HikvisionLprService : IHikvisionLprService, ILprDevice, ISingletonD
         if (plateInfoEx.byColor == null || plateInfoEx.byColor.Length == 0)
             return null;
 
-        var colorValue = plateInfoEx.byColor[0];
+        var colorValue = (int)plateInfoEx.byColor[0];
         if (!Enum.IsDefined(typeof(HikvisionPlateColorType), colorValue))
             return null;
